+++ changelog.md
@@ -1,4 +1,5 @@
 ## 1.1
+* Added check for length of input number
 * Fixed a grammatical error with bulls and cows plural output
 * Removed debug output that gave out the guessed number so game makes sense now